requirements.txt
# JobBot dependencies
# pip install -r requirements.txt

openai>=1.30
playwright>=1.44
httpx>=0.27
beautifulsoup4>=4.12
lxml>=5.2          # XML parser for RSS feeds
pypdf>=5.4        # Extract text from existing PDF resumes for scoring
python-dotenv>=1.0
